+++ drivers/stepper/Kconfig.h_bridge
@@ -2,7 +2,7 @@
 # SPDX-FileCopyrightText: Copyright (c) 2024 Jilay Sandeep Pandya
 # SPDX-License-Identifier: Apache-2.0
 
-config GPIO_STEPPER
+config H_BRIDGE_STEPPER
 	bool "Activate driver for gpio stepper control"
-	depends on DT_HAS_ZEPHYR_GPIO_STEPPER_ENABLED
+	depends on DT_HAS_ZEPHYR_H_BRIDGE_STEPPER_ENABLED
 	default y
