+++ tests/drivers/stepper/stepper_api/boards/native_sim_zephyr_h_bridge_stepper.overlay
@@ -7,18 +7,19 @@
 
 / {
 	aliases {
-		stepper =  &zephyr_gpio_stepper;
+		stepper =  &zephyr_h_bridge_stepper;
 	};
 };
 
 / {
-	zephyr_gpio_stepper: zephyr_gpio_stepper {
-		compatible = "zephyr,gpio-stepper";
+	zephyr_h_bridge_stepper: zephyr_h_bridge_stepper {
+		compatible = "zephyr,h-bridge-stepper";
 		status = "okay";
 		micro-step-res = <1>;
+		en-gpios = <&gpio1 0 0>;
 		gpios = <&gpio1 0 0>,
-				<&gpio2 0 0>,
-				<&gpio3 0 0>,
-				<&gpio4 0 0>;
+			<&gpio2 0 0>,
+			<&gpio3 0 0>,
+			<&gpio4 0 0>;
 	};
 };
