+++ tests/drivers/stepper/stepper_api/testcase.yaml
@@ -63,8 +63,8 @@ tests:
       - native_sim/native/64
   drivers.stepper.stepper_api.zephyr_gpio_stepper:
     extra_args:
-      - platform:native_sim/native/64:DTC_OVERLAY_FILE="boards/native_sim_zephyr_gpio_stepper.overlay"
-      - platform:qemu_x86_64/atom:DTC_OVERLAY_FILE="boards/qemu_x86_64_zephyr_gpio_stepper.overlay"
+      - platform:native_sim/native/64:DTC_OVERLAY_FILE="boards/native_sim_zephyr_h_bridge_stepper.overlay"
+      - platform:qemu_x86_64/atom:DTC_OVERLAY_FILE="boards/qemu_x86_64_zephyr_h_bridge_stepper.overlay"
     extra_configs:
       - CONFIG_GPIO=y
     platform_allow:
